Update System.IdentityModel.Tokens.Jwt.yaml
Removing this one since the repo history is kind of confusing and I am having an issue finding the commit for this version. There is a release4x branch that has the other 4.0.x.x versions, but not 4.0.4.403061554, and has a different license. Submitted a separate PR for this one so I can look into is more without holding up the other changes, which are fine.