add MIT license for version 1.0 of rubygems/-/hoe-gemspec
hoe-gemspec 1.0.0 is the only released version and was released 09-03-2010.  The MIT LICENSE file and the reference to MIT in the README were added in 2018. It appears that the authors intend for the license to be MIT but are not going to release the gem again just to update the license.